C 모듈의 encode/decode 함수 반환형을 ptrdiff_t로 정정 (64비트 환경)
64비트 환경을 위해 encode/decode 함수 반환형을 int 대신 ptrdiff_t로 정정

+++ src/hypua2jamo-c/hypua2jamo.h
@@ -1,35 +1,37 @@
 #ifndef __hypua2jamo_h__
 #define __hypua2jamo_h__
 
+#include <stddef.h>
+
 /* encoder */
 
 int hypua_c2d_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_c2d_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_c2d_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_c2d_ucs2_calcsize(const unsigned short *src, int srclen);
-int hypua_c2d_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_c2d_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
 int hypua_p2jc_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_p2jc_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_p2jc_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_p2jc_ucs2_calcsize(const unsigned short *src, int srclen);
-int hypua_p2jc_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_p2jc_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
 int hypua_p2jd_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_p2jd_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_p2jd_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_p2jd_ucs2_calcsize(const unsigned short *src, int srclen);
-int hypua_p2jd_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_p2jd_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
 
 /* decoder */
 
 int hypua_d2c_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_d2c_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_d2c_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_d2c_ucs2_calcsize(const unsigned short *src, int srclen);
-int hypua_d2c_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_d2c_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
 int hypua_jc2p_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_jc2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_jc2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_jc2p_ucs2_calcsize(const unsigned short *src, int srclen);
-int hypua_jc2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_jc2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
 int hypua_jd2p_ucs4_calcsize(const unsigned int *src, int srclen);
-int hypua_jd2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
+ptrdiff_t hypua_jd2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
 int hypua_jd2p_calcsize_ucs2(const unsigned short *src, int srclen);
-int hypua_jd2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
+ptrdiff_t hypua_jd2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
 
 /* decoder generic */
 
@@ -48,19 +50,19 @@ int hypua_decoder_setstate(void *decoder, int state);
 int hypua_decoder_calcsize_ucs2(void *decoder, void *src, int srclen);
 int hypua_decoder_calcsize_ucs4(void *decoder, void *src, int srclen);
 int hypua_decoder_calcsize_flush(void *t);
-int hypua_decoder_decode_ucs2(
+ptrdiff_t hypua_decoder_decode_ucs2(
 		void *decoder,
 		void *src,
 		int srclen,
 		void *dst
 );
-int hypua_decoder_decode_ucs4(
+ptrdiff_t hypua_decoder_decode_ucs4(
 		void *decoder,
 		void *src,
 		int srclen,
 		void *dst
 );
-int hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
-int hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
+ptrdiff_t hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
+ptrdiff_t hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
 
 #endif
